Fix version errors

+++ setup.sh
@@ -13,7 +13,7 @@ then
     re=".* version ([0-9])\.([0-9]).*"
     if [[ "$clang_version" =~ $re ]] 
     then
-        if [ "${BASH_REMATCH[1]}" -eq "6" ]
+        if [ "${BASH_REMATCH[1]}" -eq "5" ]
         then
             echo "Found clang-${BASH_REMATCH[1]}.${BASH_REMATCH[2]}"
             INSTALL_CLANG=0
@@ -26,7 +26,7 @@ llvm_version=$(llvm-config --version)
 re="([0-9])\.([0-9]).*"
 if [[ "$llvm_version" =~ $re ]] 
 then
-    if [ "${BASH_REMATCH[1]}" -eq "6" ]
+    if [ "${BASH_REMATCH[1]}" -eq "5" ]
     then
         echo "Found llvm-${BASH_REMATCH[1]}.${BASH_REMATCH[2]}"
         INSTALL_LLVM=0
@@ -50,18 +50,21 @@ fi
 
 if [ -f "$INSTALL_CLANG" ]
 then
-    echo "Install clang locally: Use http://releases.llvm.org/download.html#6.0.1 for your platform"
+    echo "Install clang locally: Use http://releases.llvm.org/download.html#5.0.2 for your platform"
 fi
 
 if [ -f "$INSTALL_LLVM" ]
 then
-    echo "Install llvm locally: Use http://releases.llvm.org/download.html#6.0.1 for your platform"
+    echo "Install llvm locally: Use http://releases.llvm.org/download.html#5.0.2 for your platform"
 fi
 
 if [ -f "$INSTALL_CUDA" ]
 then
     echo "Install cuda locally: Use https://developer.nvidia.com/cuda-90-download-archive for your platform"
 fi
+
+export CC=/usr/bin/clang
+export CXX=/usr/bin/clang++
 
 # Init submodules
 git submodule update --init
@@ -70,7 +73,7 @@ git submodule update --init
 cd third_party/Pangolin
 mkdir build
 cd build
-cmake .. -DCMAKE_INSTALL_PREFIX=.. -DCMAKE_BUILD_TYPE=Release -DCC=/usr/bin/clang -DCXX=/usr/bin/clang++
+cmake .. -DCMAKE_INSTALL_PREFIX=.. -DCMAKE_BUILD_TYPE=Release
 make -j8 install
 cd $SOURCE_DIR
 
@@ -78,18 +81,21 @@ cd $SOURCE_DIR
 cd third_party/cuda-image
 mkdir build
 cd build
-cmake .. -DCMAKE_INSTALL_PREFIX=.. -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED=ON -DPANGOLIN_DIR=$SOURCE_DIR/third_party/Pangolin -DEIGEN_DIR=$SOURCE_DIR/third_party/Eigen -DEIGEN3_INCLUDE_DIR=$SOURCE_DIR/third_party/Eigen/include -DCC=/usr/bin/clang -DCXX=/usr/bin/clang++
+cmake .. -DCMAKE_INSTALL_PREFIX=.. -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED=ON -DPANGOLIN_DIR=$SOURCE_DIR/third_party/Pangolin -DEIGEN3_INCLUDE_DIR=$SOURCE_DIR/third_party/Eigen/include
 make -j8 install
 cd $SOURCE_DIR
+
 
 # Setup Opt
 cd third_party
 
 ## Download terra
-wget -N https://github.com/zdevito/terra/releases/download/release-1.0.0-beta1/terra-Linux-x86_64-2e2032e.zip
-unzip -q -o terra-Linux-x86_64-2e2032e.zip 
+if [ ! -d "terra-Linux-x86_64-2e2032e" ] 
+    wget -N https://github.com/zdevito/terra/releases/download/release-1.0.0-beta1/terra-Linux-x86_64-2e2032e.zip
+    unzip -q -o terra-Linux-x86_64-2e2032e.zip 
+    rm terra-Linux-x86_64-2e2032e.zip
+fi
 ln -s terra-Linux-x86_64-2e2032e terra
-rm terra-Linux-x86_64-2e2032e.zip
 
 # Setup Opt
 cd Opt/API
